feat: basic fastapi setup gitingest trigger endpoint (#10)
* setup: set up this project + install packages

* minor_feat: create entry fastapi file for this project

* feat: first version of gitingest script

* feat: use BackgroundTasks to not let user wait for gitingest process

I might find a better way to do this so that users know when the process is done too

* feat: use subprocess and run command line of gitingest

When I run the library, i got error with asyncio process
so i switch to use subprocess so that fastapi and gitingest
don't conflict each other.

* ignore processed repo content

* feat: add a check for processed repo before ingesting it

I have 2 cases to handle any incoming repo request, so that i
don't have to process a repo twice. I also remove BackgroundTasks
because it doesn't let the users know what's the progress of the
operation

* minor_feat: use 'repository' as marker for summary block

and append summary to the file, instead of reading and rewriting
the whole file (not so efficient)

* feat: add more extensive error handling and messages

* feat: return 2 values for better error handling